backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart
requests
agentscope
tree-sitter>=0.21.0
tree-sitter-python
tree-sitter-typescript
tree-sitter-javascript
chromadb
networkx>=3.0.0
pydantic>=2.0.0
pytest
GitPython
